login logout AI

+++ src/main/java/org/pbl4/pbl4_be/models/GameMove.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class GameMove {
     private int row;
     private int col;
-    private byte nthMove;
+    private short nthMove;
     private boolean isWin;
 
     public GameMove() {
